22.051 (Housekeeping Update)
Oh wow, this was...a long time coming! Well, this is probably gonna be the last of the "old-school" me - Rare's learned how to actually code in Python since two years ago, and he's working on a complete rewrite of me that's gonna hopefully be MUCH easier to maintain going forward! But in the meantime, it's only right I give you one last hurrah with the long-awaited Housekeeping Update, hey?~ Enjoy!
--
!! IMPORTANT !!
Dependencies have been updated. View README.md for more information.

CHANGELOG
Added and Improved:
+ Moved internal settings and the colorset to separate .ini files - you won't need to open the main script anymore!
+ Replaced the chost.dat file with a .ini file, that way you can see the User ID of the host of any given channel. This will be auto-generated by the script on first launch.
+ Tweaked some inconsistencies. Hopefully.
+ Other changes made incrementally without anyone noticing.

Removed:
- The "lock" and "unlock" commands have been removed due to inconsistent performance and a habit of locking up the script.